Moves Discord Command parsing to `Load` phase from `init` phase
The skript parser calls `init` when parsing the structures for the first time, but before the structures are actually loaded. So because `CommandFactory.add()` uses `SkriptUtils.loadCode(trigger, CommandEvent.class)` when creating the CommandObject it causes the trigger code to be loaded before any function references have been registered (as that happens in the `preload` phase.

Fixes https://github.com/DiSkyOrg/DiSky/issues/173